+++ packages/axios/src/index.ts
@@ -1,21 +1,38 @@
-import { computed } from "@vue/composition-api";
-import axios, { AxiosRequestConfig } from "axios";
-import { usePromise } from "@vue-composable/core";
+import { computed, Ref } from "@vue/composition-api";
+import axios, { AxiosRequestConfig, AxiosResponse, AxiosInstance } from "axios";
+import { usePromise, PromiseResultFactory } from "@vue-composable/core";
 
 /* istanbul ignore next  */
 const _axios = axios || (globalThis && (globalThis as any).axios);
 
-export function useAxios<TData = any>(config?: AxiosRequestConfig) {
+interface AxiosReturn<TData> extends PromiseResultFactory<Promise<AxiosResponse<TData>>, [AxiosRequestConfig]> {
+  readonly client: Ref<Readonly<AxiosInstance>>;
+  readonly data: Ref<TData | null>;
+  readonly status: Ref<number | null>;
+  readonly statusText: Ref<string | null>;
+}
+
+// TODO replace by shared project
+const isObject = (d: any): d is Object => typeof d === 'object';
+const isString = (d: any): d is String => typeof d === 'string';
+const isBoolean = (d: any): d is Boolean => typeof d === 'boolean';
+
+export function useAxios<TData = any>(url: string, config?: AxiosRequestConfig, throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(config?: AxiosRequestConfig, throwException?: boolean): AxiosReturn<TData>;
+export function useAxios<TData = any>(configUrl?: AxiosRequestConfig | string, configThrowException?: AxiosRequestConfig | boolean, throwException = false): AxiosReturn<TData> {
   /* istanbul ignore next  */
   __DEV__ &&
     !_axios &&
     console.warn(`[axios] not installed, please install it`);
 
+  const config = !isString(configUrl) ? configUrl : isObject(configThrowException) ? configThrowException as AxiosRequestConfig : undefined;
+  throwException = isBoolean(configThrowException) ? configThrowException : throwException;
+
   const axiosClient = _axios.create(config);
   const client = computed(() => axiosClient);
   const use = usePromise(async (request: AxiosRequestConfig) => {
-    return axiosClient.request(request);
-  });
+    return axiosClient.request<any, AxiosResponse<TData>>(request);
+  }, throwException);
 
   const data = computed<TData>(
     () =>
@@ -25,7 +42,7 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
         use.error.value.response.data) ||
       null
   );
-  const status = computed(
+  const status = computed<number>(
     () =>
       (use.result.value && use.result.value.status) ||
       (use.error.value &&
@@ -33,7 +50,7 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
         use.error.value.response.status) ||
       null
   );
-  const statusText = computed(
+  const statusText = computed<string>(
     () =>
       (use.result.value && use.result.value.statusText) ||
       (use.error.value &&
@@ -41,6 +58,13 @@ export function useAxios<TData = any>(config?: AxiosRequestConfig) {
         use.error.value.response.statusText) ||
       null
   );
+
+  // if url provided in the config, execute it straight away
+  if (typeof configUrl === 'string') {
+    use.exec({ url: configUrl });
+  } else if (config && config.url) {
+    use.exec(config)
+  }
 
   return {
     ...use,