+++ packages/core/src/utils.ts
@@ -19,10 +19,10 @@ export function wrap<T>(o: RefTyped<T>): Ref<T> {
 export const isArray = Array.isArray;
 export const isFunction = (val: unknown): val is Function =>
   typeof val === "function";
-// export const isString = (val: unknown): val is string =>
-//   typeof val === "string";
-// export const isSymbol = (val: unknown): val is symbol =>
-//   typeof val === "symbol";
+export const isString = (val: unknown): val is string =>
+  typeof val === "string";
+export const isSymbol = (val: unknown): val is symbol =>
+  typeof val === "symbol";
 
 export const isDate = (val: unknown): val is Date =>
   isObject(val) && isFunction(val.getTime);