+++ CHANGELOG.md
@@ -3,13 +3,15 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-<!-- 
+
 
 ## [Unreleased]
 
 --- 
 
--->
+### Added
+- `usePromise` - added new argument `throwException`, default: `false`, if true it will return the exception when using `exec`
+
 
 ## 1.0.0-dev.2